Fix a few edge cases of diffing column constraints
1. Column constraints were being removed after being added, resulting
in deleting the constraint in any case where it changed.
2. Default timestamp with time zone column constraints were being
compared too naively. Now we at least attempt to parse the formats
produced by postgresql-simple and postgres itself, and if we can,
compare the resulting UTCTimes to see if anything needs to change.
To aid this, we also set the transaction-local time zone to UTC before
getting the schema, so that absolute timestamps in default constraints
are reported in UTC time.
3. Default enumeration values were getting type-annotated by postgres
but not by the Haskell side, we try to look for this and avoid changing
things if it's the only difference.